refactor :  route 가 아닌 폴더의 네이밍을 수정합니다.
